Break out the ReplayableStream
